chore: fix the lack of update of mapTrees, since we were splaying        the trees and forgetting to update the mapTrees.

+++ src/decremental_minimum_spanning_forest/helpful_classes/minHeap.hpp
@@ -68,7 +68,7 @@ class MinHeap {
                 return;
             }
             heap.push_back({v, weight});
-            int index = heap.size() - 1;
+            long unsigned int index = heap.size() - 1;
             position[v] = index;
             heapifyUp(index);
         }
@@ -77,7 +77,7 @@ class MinHeap {
         void remove(Key v) {
             if (!position.count(v)) return;
             
-            int index = position[v];
+            long unsigned int index = position[v];
             position.erase(v);
             
             if (index == heap.size() - 1) {
